+++ RedPandaIDE/settings.cpp
@@ -1669,7 +1669,7 @@ Settings::CompilerSet::CompilerSet():
     mAutoAddCharsetParams{false},
     mExecCharset{ENCODING_SYSTEM_DEFAULT},
     mStaticLink{false},
-    mMaxObjectSize{0},
+    mMaxFrameSize{0},
     mWarnLargeObject{false},
     mPreprocessingSuffix{DEFAULT_PREPROCESSING_SUFFIX},
     mCompilationProperSuffix{DEFAULT_COMPILATION_SUFFIX},
@@ -1685,7 +1685,7 @@ Settings::CompilerSet::CompilerSet(const QString& compilerFolder, const QString&
     mAutoAddCharsetParams{true},
     mExecCharset{ENCODING_SYSTEM_DEFAULT},
     mStaticLink{true},
-    mMaxObjectSize{0},
+    mMaxFrameSize{0},
     mWarnLargeObject{false},
     mPreprocessingSuffix{DEFAULT_PREPROCESSING_SUFFIX},
     mCompilationProperSuffix{DEFAULT_COMPILATION_SUFFIX},
@@ -1742,7 +1742,7 @@ Settings::CompilerSet::CompilerSet(const Settings::CompilerSet &set):
     mAutoAddCharsetParams{set.mAutoAddCharsetParams},
     mExecCharset{set.mExecCharset},
     mStaticLink{set.mStaticLink},
-    mMaxObjectSize{set.mMaxObjectSize},
+    mMaxFrameSize{set.mMaxFrameSize},
     mWarnLargeObject{set.mWarnLargeObject},
 
     mPreprocessingSuffix{set.mPreprocessingSuffix},
@@ -2569,24 +2569,24 @@ QByteArray Settings::CompilerSet::getCompilerOutput(const QString &binDir, const
     return result.trimmed();
 }
 
-bool Settings::CompilerSet::warnLargeObject() const
+bool Settings::CompilerSet::warnLargeFrame() const
 {
     return mWarnLargeObject;
 }
 
-void Settings::CompilerSet::setWarnLargeObject(bool newWarnLargeObject)
+void Settings::CompilerSet::setWarnLargeFrame(bool newWarnLargeObject)
 {
     mWarnLargeObject = newWarnLargeObject;
 }
 
-double Settings::CompilerSet::maxObjectSize() const
+double Settings::CompilerSet::maxFrameSize() const
 {
-    return mMaxObjectSize;
+    return mMaxFrameSize;
 }
 
-void Settings::CompilerSet::setMaxObjectSize(double maxObjectSize)
+void Settings::CompilerSet::setMaxFrameSize(double maxFrameSize)
 {
-    mMaxObjectSize = maxObjectSize;
+    mMaxFrameSize = maxFrameSize;
 }
 
 Settings::CompilerSet::CompilationStage Settings::CompilerSet::compilationStage() const
@@ -2758,8 +2758,8 @@ static void setReleaseOptions(Settings::PCompilerSet pSet) {
     pSet->setCompileOption(LINK_CMD_OPT_STRIP_EXE, COMPILER_OPTION_ON);
     pSet->setCompileOption(CC_CMD_OPT_USE_PIPE, COMPILER_OPTION_ON);
     pSet->setStaticLink(true);
-    pSet->setMaxObjectSize(2); //default stack size of gcc is 2MB
-    pSet->setWarnLargeObject(false);
+    pSet->setMaxFrameSize(2); //default stack size of gcc is 2MB
+    pSet->setWarnLargeFrame(false);
 }
 
 static void setDebugOptions(Settings::PCompilerSet pSet, bool enableAsan = false) {
@@ -2779,8 +2779,8 @@ static void setDebugOptions(Settings::PCompilerSet pSet, bool enableAsan = false
     pSet->setCompileOption(CC_CMD_OPT_STACK_PROTECTOR, "-strong");
     pSet->setStaticLink(false);
 
-    pSet->setMaxObjectSize(2); //default stack size of gcc is 2MB
-    pSet->setWarnLargeObject(true);
+    pSet->setMaxFrameSize(2); //default stack size of gcc is 2MB
+    pSet->setWarnLargeFrame(true);
 }
 
 bool Settings::CompilerSets::addSets(const QString &folder, const QString& c_prog) {
@@ -3122,8 +3122,9 @@ void Settings::CompilerSets::saveSet(int index)
     mSettings->mSettings.setValue("AddCharset", pSet->autoAddCharsetParams());
     mSettings->mSettings.setValue("StaticLink", pSet->staticLink());
     mSettings->mSettings.setValue("ExecCharset", pSet->execCharset());
-    mSettings->mSettings.setValue("WarnLargeObject",pSet->warnLargeObject());
-    mSettings->mSettings.setValue("MaxObjectSize",pSet->maxObjectSize());
+    mSettings->mSettings.setValue("WarnLargeFrame",pSet->warnLargeFrame());
+    mSettings->mSettings.setValue("MaxFrameSize",pSet->maxFrameSize());
+
 
     mSettings->mSettings.setValue("preprocessingSuffix", pSet->preprocessingSuffix());
     mSettings->mSettings.setValue("compilationProperSuffix", pSet->compilationProperSuffix());
@@ -3207,8 +3208,8 @@ Settings::PCompilerSet Settings::CompilerSets::loadSet(int index)
     pSet->setCustomLinkParams(mSettings->mSettings.value("customLinkParams").toString());
     pSet->setAutoAddCharsetParams(mSettings->mSettings.value("AddCharset", true).toBool());
     pSet->setStaticLink(mSettings->mSettings.value("StaticLink", false).toBool());
-    pSet->setMaxObjectSize(mSettings->mSettings.value("MaxObjectSize", 2).toDouble());
-    pSet->setWarnLargeObject(mSettings->mSettings.value("WarnLargeObject", false).toBool());
+    pSet->setMaxFrameSize(mSettings->mSettings.value("MaxFrameSize", 1).toDouble());
+    pSet->setWarnLargeFrame(mSettings->mSettings.value("WarnLargeFrame", false).toBool());
 
     pSet->setExecCharset(mSettings->mSettings.value("ExecCharset", ENCODING_SYSTEM_DEFAULT).toString());
     if (pSet->execCharset().isEmpty()) {